A sanity test running successfully with tb and model being at place

+++ C_code_source_FPGA/IIR_filter.h
@@ -9,31 +9,33 @@
 #define SRC_IIR_FILTER_H_
 
 #include "stdint.h"
+#include <stdio.h>
+#include <math.h>
 
 #define INPUT_BUFFER_LENGTH 3
 #define OUTPUT_BUFFER_LENGTH 2
 
 
 typedef struct {
-	short input_buffer[INPUT_BUFFER_LENGTH];
+	int input_buffer[INPUT_BUFFER_LENGTH];
 	uint8_t input_buffer_index;
 
-	short output_buffer[OUTPUT_BUFFER_LENGTH];
+	int output_buffer[OUTPUT_BUFFER_LENGTH];
 	uint8_t output_buffer_index;
 
-	short output;
+	int output;
 
 }filter;
 
 typedef struct{
-	float input_coeff[INPUT_BUFFER_LENGTH];
-	float output_coeff[OUTPUT_BUFFER_LENGTH];
+	double input_coeff[INPUT_BUFFER_LENGTH];
+	double output_coeff[OUTPUT_BUFFER_LENGTH];
 
 	const short fabric_input_coeff[INPUT_BUFFER_LENGTH];
 	const short fabric_output_coeff[OUTPUT_BUFFER_LENGTH];  
 }coeff_struct;
 
 void init_filter(filter *);
-short compute_filter_output(filter* , short, float[], float[] );
+int compute_filter_output(filter* , int, double[], double[] );
 
 #endif /* SRC_IIR_FILTER_H_ */
